% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationR.R
\name{bisection}
\alias{bisection}
\title{Bisection Method for Root Finding}
\usage{
bisection(a, b, x, tol = 1e-06, max_iter = 1000L)
}
\arguments{
\item{a}{Lower bound of interval}

\item{b}{Upper bound of interval}

\item{x}{Data vector}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}
}
\value{
List containing the root and number of iterations
}
\description{
Bisection Method for Root Finding
}
