% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{brents}
\alias{brents}
\title{Brent's Method for Optimization}
\usage{
brents(f, a, b, iteration = 20, tol = 1e-05)
}
\arguments{
\item{f}{A function that takes one argument x that is to be optimized
Note that the function has to be continuous over the interval stated and must have a Lagrange interpolating polynomial of 2 or more}

\item{a}{Known as the contrapoint such that f(a) and f(b) need to contain different signs to each other and the interval contains the solution.}

\item{b}{The current iterate of the algorithm that is the current guess for the rootof the function.}

\item{iteration}{Number of iterations that the algorithm should run. Default is set at 20.}

\item{tol}{A tolerance level of which is used to show that the function is convergingtowards the optimal solution. Default is set at 1e-05}
}
\description{
Brent's Method for Optimization
}
\examples{
> brents(g, 2, 5, 16)
 Number of iterations 16
 Root 4.472153
}
