% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationR.R
\name{cauchy_log_likelihood}
\alias{cauchy_log_likelihood}
\alias{cauchy_score}
\alias{cauchy_observed_info}
\title{Cauchy Distribution Log-Likelihood and Derivatives}
\usage{
cauchy_log_likelihood(theta, x)

cauchy_score(theta, x)

cauchy_observed_info(theta, x)
}
\arguments{
\item{theta}{Location parameter}

\item{x}{Numeric vector of data}
}
\value{
Function value
}
\description{
Helper functions for Cauchy distribution optimization
}
