% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationR.R
\name{fisher_scoring}
\alias{fisher_scoring}
\title{Fisher Scoring Method for Root Finding}
\usage{
fisher_scoring(theta0, x, tol = 1e-06, max_iter = 1000L)
}
\arguments{
\item{theta0}{Initial guess}

\item{x}{Data vector}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}
}
\value{
List containing the root and number of iterations
}
\description{
Fisher Scoring Method for Root Finding
}
