% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoldenSectionSearch.R
\name{Golden_Section_Search}
\alias{Golden_Section_Search}
\title{Finds the minimum point of a function}
\usage{
Golden_Section_Search(f, ub, lb, Tolerance)
}
\arguments{
\item{f}{function}

\item{ub}{Upper Bound}

\item{lb}{Lower Bound}

\item{Tolerance}{Accuracy of answer}
}
\description{
Finds the minimum point of a function
}
