% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{graph}
\alias{graph}
\title{Plot the graph of a function}
\usage{
graph(x, fun)
}
\arguments{
\item{x}{\code{vector} points in the domain to evaluate the function}

\item{fun}{\code{function} to be evaluated}
}
\description{
Plot the graph of a function
}

