% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{gss}
\alias{gss}
\title{Golden-section search function}
\usage{
gss(f, a, b, it = 10, tol = 1e-05)
}
\arguments{
\item{f}{A function that takes one argument x that is to be optimized.
The function must be strictly unimodal on the interval [a, b], if there is a unique value in [a,b] such that  is the minimum of f is on [a,b]}

\item{a}{The lower bound of the function, in which the optimal value is larger than. This value must also be real and scalar.}

\item{b}{The upper bound of the function, in which the optimal value is smaller than. This value must also be real and scalar.}

\item{it}{Number of iterations that the algorithm should run. Default is set at 10.}

\item{tol}{A tolerance level of which is used to show that the function is convergingtowards the optimal solution. Set at 1e-05}
}
\description{
Golden-section search function
}
\examples{
> gss(f, 1, 5)
 Interation number 10
 The subset interval 1.989217 1.996894
 The estimated minimizer 1.993056
}
