% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationR.R
\name{multi_newton}
\alias{multi_newton}
\title{Multivaribale Newton-Raphson Method for Root Finding}
\usage{
multi_newton(X, y, tol = 1e-06, max_iter = 1000L)
}
\arguments{
\item{X}{Data matrix}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}

\item{beta}{Initial guess}
}
\value{
List containing the root and number of iterations
}
\description{
Multivaribale Newton-Raphson Method for Root Finding
}
