% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationR.R
\name{newton}
\alias{newton}
\title{Newton-Raphson Method for Root Finding}
\usage{
newton(theta0, x, tol = 1e-06, max_iter = 1000L)
}
\arguments{
\item{theta0}{Initial guess}

\item{x}{Data vector}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}
}
\value{
List containing the root and number of iterations
}
\description{
Newton-Raphson Method for Root Finding
}
