% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Newton Raphson.R
\name{newton_raphson}
\alias{newton_raphson}
\title{Finds the minimum point of a function}
\usage{
newton_raphson(f, a, tolerance, n = 1000)
}
\arguments{
\item{f}{function}

\item{a}{Initial starting point}

\item{tolerance}{accuracy of the answer}

\item{n}{Number of iterations}
}
\description{
Finds the minimum point of a function
}
