% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotGraph.R
\name{plotgraph}
\alias{plotgraph}
\title{Finds the minimum point of a function}
\usage{
plotgraph(f, a, b)
}
\arguments{
\item{f}{Function}

\item{a}{Minimum point of the graph}

\item{b}{Maximum point of the graph}

\item{x}{Variable of the function}
}
\description{
Finds the minimum point of a function
}
