% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Secant.R
\name{Secant}
\alias{Secant}
\title{Finds the minimum point of a function}
\usage{
Secant(f, x0, x1, tolerance, n = 1000)
}
\arguments{
\item{f}{function}

\item{x0}{First initial point where x0 < x1}

\item{x1}{Second initial point where x0 < x1}
}
\description{
Finds the minimum point of a function
}
