% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unifisher}
\alias{unifisher}
\title{Uses the Fisher scoring method to find the root of a univariate likelihood function, denoted l.}
\usage{
unifisher(lp, fisherInfo, n, a, eps, maxIter)
}
\arguments{
\item{lp}{<func> First derivative of l, function of only the parameter.}

\item{fisherInfo}{<func> Fisher information, function of only the parameter.}

\item{n}{<int> Sample size.}

\item{a}{<double> Initial guess for parameter.}

\item{eps}{<double> Tolerance used to classify convergence.}

\item{maxIter}{<int> Maximum number of iterations to perform before stopping.}
}
\value{
<list> Estimate for parameter and the number of iterations required for convergence.
}
\description{
Uses the Fisher scoring method to find the root of a univariate likelihood function, denoted l.
}
