% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{uninewton}
\alias{uninewton}
\title{Uses the Newton method to find the root of a univariate function, denoted g.}
\usage{
uninewton(gp, gpp, a, eps, maxIter)
}
\arguments{
\item{gp}{<func> First derivative of g.}

\item{gpp}{<func> Second derivative of g.}

\item{a}{<double> Initial guess for the root.}

\item{eps}{<double> Tolerance used to classify convergence.}

\item{maxIter}{<int> Maximum number of iterations to perform before stopping.}
}
\value{
<list> Root of g and the number of iterations required for convergence.
}
\description{
Uses the Newton method to find the root of a univariate function, denoted g.
}
