% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Optimization.R
\name{newtons}
\alias{newtons}
\title{Newton's Method}
\usage{
newtons(f, x0, it = 10, tol = 1e-07)
}
\arguments{
\item{f}{Function f that takes one variable x}

\item{x0}{The intial guess for the root of the function}

\item{it}{Number of iterations the algorithm should run, default set at 10}

\item{tol}{Tolerance level, deafukt set at 1e-07}
}
\value{
root of the function and the number of iterations ran
}
\description{
Newton's Method
}
\examples{
z <- function(x){
return(cos(x) - x**3 )
}
> newtons(z, 0.5)
Number of iteration 6
Solution 0.865474

}
