% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OptimizationR.R
\name{secant}
\alias{secant}
\title{Secant Method for Root Finding}
\usage{
secant(theta0, theta1, x, tol = 1e-06, max_iter = 1000L)
}
\arguments{
\item{theta0}{First Initial guess}

\item{theta1}{Second Initial guess}

\item{x}{Data vector}

\item{tol}{Tolerance for convergence}

\item{max_iter}{Maximum number of iterations}
}
\value{
List containing the root and number of iterations
}
\description{
Secant Method for Root Finding
}
